Refactor Test Code
