+++ src/test/lib/CMakeLists.txt
@@ -3,10 +3,13 @@
 
 set(SOURCES
     ApiTest.cpp
+    BasicTest.cpp
+    DatagramTest.cpp
+    DataTest.cpp
     DrillDescriptor.cpp
     EventTest.cpp
+    HandshakeTest.cpp
     QuicDrill.cpp
-    QuicTest.cpp
     TestConnection.cpp
     TestListener.cpp
     TestStream.cpp
