+++ src/test/lib/DataTest.cpp
@@ -5,73 +5,11 @@
 
 Abstract:
 
-    MsQuic API Unittest
+    Tests various features related to the data path.
 
 --*/
 
 #include "precomp.h"
-
-//#define QUIC_TEST_DISABLE_DNS 1
-
-#define OLD_SUPPORTED_VERSION       QUIC_VERSION_1_MS_H
-#define LATEST_SUPPORTED_VERSION    QUIC_VERSION_LATEST_H
-
-const uint16_t TestUdpPortBase = 0x8000;
-
-const QuicAddr GetLocalAddr(int Family)
-{
-    return Family == 4 ? QuicAddr(AF_INET, true) : QuicAddr(AF_INET6, true);
-}
-
-void QuicTestInitialize()
-{
-#ifdef QUIC_NO_ENCRYPTION
-    uint8_t Disabled = FALSE;
-    if (QUIC_FAILED(
-        MsQuic->SetParam(
-            nullptr,
-            QUIC_PARAM_LEVEL_GLOBAL,
-            QUIC_PARAM_GLOBAL_ENCRYPTION,
-            sizeof(Disabled),
-            &Disabled))) {
-        QuicTraceLogError(
-            TestDisableEncryptionFailure,
-            "[test] Disabling encryption failed");
-    }
-#endif
-}
-
-void QuicTestCleanup()
-{
-}
-
-struct TestScopeLogger
-{
-    const char* Name;
-    TestScopeLogger(const char* name) : Name(name) {
-        QuicTraceLogInfo(
-            TestScopeEntry,
-            "[test]---> %s",
-            Name);
-    }
-    ~TestScopeLogger() {
-        QuicTraceLogInfo(
-            TestScopeExit,
-            "[test]<--- %s",
-            Name);
-    }
-};
-
-_Function_class_(NEW_CONNECTION_CALLBACK)
-static
-void
-ListenerDoNothingCallback(
-    _In_ TestListener* /* Listener */,
-    _In_ HQUIC /* ConnectionHandle */
-    )
-{
-    TEST_FAILURE("This callback should never be called!");
-}
 
 _Function_class_(NEW_STREAM_CALLBACK)
 static
@@ -85,215 +23,6 @@ ConnectionDoNothingCallback(
     TEST_FAILURE("This callback should never be called!");
 }
 
-void QuicTestCreateListener()
-{
-    MsQuicSession Session;
-    TEST_TRUE(Session.IsValid());
-
-    {
-        TestListener Listener(Session.Handle, ListenerDoNothingCallback);
-        TEST_TRUE(Listener.IsValid());
-    }
-}
-
-void QuicTestStartListener()
-{
-    MsQuicSession Session;
-    TEST_TRUE(Session.IsValid());
-
-    {
-        TestListener Listener(Session.Handle, ListenerDoNothingCallback);
-        TEST_TRUE(Listener.IsValid());
-        TEST_QUIC_SUCCEEDED(Listener.Start());
-    }
-
-    {
-        TestListener Listener(Session.Handle, ListenerDoNothingCallback);
-        TEST_TRUE(Listener.IsValid());
-        QuicAddr LocalAddress(AF_UNSPEC);
-        TEST_QUIC_SUCCEEDED(Listener.Start(&LocalAddress.SockAddr));
-    }
-}
-
-void QuicTestStartListenerMultiAlpns()
-{
-    MsQuicSession Session("MsQuicTest1", "MsQuicTest2");
-    TEST_TRUE(Session.IsValid());
-
-    {
-        TestListener Listener(Session.Handle, ListenerDoNothingCallback);
-        TEST_TRUE(Listener.IsValid());
-        TEST_QUIC_SUCCEEDED(Listener.Start());
-    }
-
-    {
-        TestListener Listener(Session.Handle, ListenerDoNothingCallback);
-        TEST_TRUE(Listener.IsValid());
-        QuicAddr LocalAddress(AF_UNSPEC);
-        TEST_QUIC_SUCCEEDED(Listener.Start(&LocalAddress.SockAddr));
-    }
-}
-
-void QuicTestStartListenerImplicit(_In_ int Family)
-{
-    MsQuicSession Session;
-    TEST_TRUE(Session.IsValid());
-
-    {
-        TestListener Listener(Session.Handle, ListenerDoNothingCallback);
-        TEST_TRUE(Listener.IsValid());
-
-        QuicAddr LocalAddress(Family == 4 ? AF_INET : AF_INET6);
-        TEST_QUIC_SUCCEEDED(Listener.Start(&LocalAddress.SockAddr));
-    }
-}
-
-void QuicTestStartTwoListeners()
-{
-    MsQuicSession Session1;
-    TEST_TRUE(Session1.IsValid());
-    MsQuicSession Session2("MsQuicTest2");
-    TEST_TRUE(Session2.IsValid());
-
-    {
-        TestListener Listener1(Session1.Handle, ListenerDoNothingCallback);
-        TEST_TRUE(Listener1.IsValid());
-        TEST_QUIC_SUCCEEDED(Listener1.Start());
-
-        QuicAddr LocalAddress;
-        TEST_QUIC_SUCCEEDED(Listener1.GetLocalAddr(LocalAddress));
-
-        TestListener Listener2(Session2.Handle, ListenerDoNothingCallback);
-        TEST_TRUE(Listener2.IsValid());
-        TEST_QUIC_SUCCEEDED(Listener2.Start(&LocalAddress.SockAddr));
-    }
-}
-
-void QuicTestStartTwoListenersSameALPN()
-{
-    MsQuicSession Session;
-    TEST_TRUE(Session.IsValid());
-    MsQuicSession Session2("MsQuicTest", "MsQuicTest2");
-    TEST_TRUE(Session2.IsValid());
-
-    {
-        //
-        // Both try to listen on the same, single ALPN
-        //
-        TestListener Listener1(Session.Handle, ListenerDoNothingCallback);
-        TEST_TRUE(Listener1.IsValid());
-        TEST_QUIC_SUCCEEDED(Listener1.Start());
-
-        QuicAddr LocalAddress;
-        TEST_QUIC_SUCCEEDED(Listener1.GetLocalAddr(LocalAddress));
-
-        TestListener Listener2(Session.Handle, ListenerDoNothingCallback);
-        TEST_TRUE(Listener2.IsValid());
-        TEST_QUIC_STATUS(
-            QUIC_STATUS_INVALID_STATE,
-            Listener2.Start(&LocalAddress.SockAddr));
-    }
-
-    {
-        //
-        // First listener on two ALPNs and second overlaps one of those.
-        //
-        TestListener Listener1(Session2.Handle, ListenerDoNothingCallback);
-        TEST_TRUE(Listener1.IsValid());
-        TEST_QUIC_SUCCEEDED(Listener1.Start());
-
-        QuicAddr LocalAddress;
-        TEST_QUIC_SUCCEEDED(Listener1.GetLocalAddr(LocalAddress));
-
-        TestListener Listener2(Session.Handle, ListenerDoNothingCallback);
-        TEST_TRUE(Listener2.IsValid());
-        TEST_QUIC_STATUS(
-            QUIC_STATUS_INVALID_STATE,
-            Listener2.Start(&LocalAddress.SockAddr));
-    }
-
-    {
-        //
-        // First listener on one ALPN and second with two (one that overlaps).
-        //
-        TestListener Listener1(Session.Handle, ListenerDoNothingCallback);
-        TEST_TRUE(Listener1.IsValid());
-        TEST_QUIC_SUCCEEDED(Listener1.Start());
-
-        QuicAddr LocalAddress;
-        TEST_QUIC_SUCCEEDED(Listener1.GetLocalAddr(LocalAddress));
-
-        TestListener Listener2(Session2.Handle, ListenerDoNothingCallback);
-        TEST_TRUE(Listener2.IsValid());
-        TEST_QUIC_STATUS(
-            QUIC_STATUS_INVALID_STATE,
-            Listener2.Start(&LocalAddress.SockAddr));
-    }
-}
-
-void QuicTestStartListenerExplicit(_In_ int Family)
-{
-    MsQuicSession Session;
-    TEST_TRUE(Session.IsValid());
-
-    {
-        TestListener Listener(Session.Handle, ListenerDoNothingCallback);
-        TEST_TRUE(Listener.IsValid());
-
-        QuicAddr LocalAddress(GetLocalAddr(Family), TestUdpPortBase);
-        QUIC_STATUS Status = QUIC_STATUS_ADDRESS_IN_USE;
-        while (Status == QUIC_STATUS_ADDRESS_IN_USE) {
-            LocalAddress.IncrementPort();
-            Status = Listener.Start(&LocalAddress.SockAddr);
-        }
-        TEST_QUIC_SUCCEEDED(Status);
-    }
-}
-
-void QuicTestCreateConnection()
-{
-    MsQuicSession Session;
-    TEST_TRUE(Session.IsValid());
-
-    {
-        TestConnection Connection(Session.Handle, ConnectionDoNothingCallback, false);
-        TEST_TRUE(Connection.IsValid());
-    }
-}
-
-void QuicTestBindConnectionImplicit(_In_ int Family)
-{
-    MsQuicSession Session;
-    TEST_TRUE(Session.IsValid());
-
-    {
-        TestConnection Connection(Session.Handle, ConnectionDoNothingCallback, false);
-        TEST_TRUE(Connection.IsValid());
-
-        QuicAddr LocalAddress(Family == 4 ? AF_INET : AF_INET6);
-        TEST_QUIC_SUCCEEDED(Connection.SetLocalAddr(LocalAddress));
-    }
-}
-
-void QuicTestBindConnectionExplicit(_In_ int Family)
-{
-    MsQuicSession Session;
-    TEST_TRUE(Session.IsValid());
-
-    {
-        TestConnection Connection(Session.Handle, ConnectionDoNothingCallback, false);
-        TEST_TRUE(Connection.IsValid());
-
-        QuicAddr LocalAddress(GetLocalAddr(Family), TestUdpPortBase);
-        QUIC_STATUS Status = QUIC_STATUS_ADDRESS_IN_USE;
-        while (Status == QUIC_STATUS_ADDRESS_IN_USE) {
-            LocalAddress.IncrementPort();
-            Status = Connection.SetLocalAddr(LocalAddress);
-        }
-        TEST_QUIC_SUCCEEDED(Status);
-    }
-}
-
 struct ServerAcceptContext {
     QUIC_EVENT NewConnectionReady;
     TestConnection** NewConnection;
@@ -305,361 +34,6 @@ struct ServerAcceptContext {
         QuicEventUninitialize(NewConnectionReady);
     }
 };
-
-_Function_class_(NEW_CONNECTION_CALLBACK)
-static
-void
-ListenerAcceptConnection(
-    _In_ TestListener* Listener,
-    _In_ HQUIC ConnectionHandle
-    )
-{
-    ServerAcceptContext* AcceptContext = (ServerAcceptContext*)Listener->Context;
-    if (AcceptContext == nullptr) { // Prime Resumption scenario.
-        auto NewConnection = new TestConnection(ConnectionHandle, ConnectionDoNothingCallback, true, true);
-        if (NewConnection == nullptr || !NewConnection->IsValid()) {
-            TEST_FAILURE("Failed to accept new TestConnection.");
-            delete NewConnection;
-            MsQuic->ConnectionClose(ConnectionHandle);
-        } else {
-            NewConnection->SetHasRandomLoss(Listener->GetHasRandomLoss());
-        }
-        return;
-    }
-    if (*AcceptContext->NewConnection != nullptr) { // Retry scenario.
-        delete *AcceptContext->NewConnection;
-        *AcceptContext->NewConnection = nullptr;
-    }
-    *AcceptContext->NewConnection = new TestConnection(ConnectionHandle, ConnectionDoNothingCallback, true);
-    if (*AcceptContext->NewConnection == nullptr || !(*AcceptContext->NewConnection)->IsValid()) {
-        TEST_FAILURE("Failed to accept new TestConnection.");
-        delete *AcceptContext->NewConnection;
-        *AcceptContext->NewConnection = nullptr;
-        MsQuic->ConnectionClose(ConnectionHandle);
-    } else {
-        (*AcceptContext->NewConnection)->SetHasRandomLoss(Listener->GetHasRandomLoss());
-    }
-    QuicEventSet(AcceptContext->NewConnectionReady);
-}
-
-struct StatelessRetryHelper
-{
-    bool DoRetry;
-    StatelessRetryHelper(bool Enabled) : DoRetry(Enabled) {
-        if (DoRetry) {
-            uint16_t value = 0;
-            TEST_QUIC_SUCCEEDED(
-                MsQuic->SetParam(
-                    nullptr,
-                    QUIC_PARAM_LEVEL_GLOBAL,
-                    QUIC_PARAM_GLOBAL_RETRY_MEMORY_PERCENT,
-                    sizeof(value),
-                    &value));
-        }
-    }
-    ~StatelessRetryHelper() {
-        if (DoRetry) {
-            uint16_t value = 65;
-            TEST_QUIC_SUCCEEDED(
-                MsQuic->SetParam(
-                    nullptr,
-                    QUIC_PARAM_LEVEL_GLOBAL,
-                    QUIC_PARAM_GLOBAL_RETRY_MEMORY_PERCENT,
-                    sizeof(value),
-                    &value));
-        }
-    }
-};
-
-#define PRIVATE_TP_TYPE   77
-#define PRIVATE_TP_LENGTH 2345
-
-struct PrivateTransportHelper : QUIC_PRIVATE_TRANSPORT_PARAMETER
-{
-    PrivateTransportHelper(bool Enabled) {
-        if (Enabled) {
-            Type = PRIVATE_TP_TYPE;
-            Length = PRIVATE_TP_LENGTH;
-            Buffer = new uint8_t[PRIVATE_TP_LENGTH];
-            TEST_TRUE(Buffer != nullptr);
-        } else {
-            Buffer = nullptr;
-        }
-    }
-    ~PrivateTransportHelper() {
-        delete [] Buffer;
-    }
-};
-
-struct RandomLossHelper
-{
-    static uint8_t LossPercentage;
-    static QUIC_TEST_DATAPATH_HOOKS DataPathFuncTable;
-    RandomLossHelper(uint8_t _LossPercentage) {
-        LossPercentage = _LossPercentage;
-        if (LossPercentage != 0) {
-            QUIC_TEST_DATAPATH_HOOKS* Value = &DataPathFuncTable;
-            TEST_QUIC_SUCCEEDED(
-                MsQuic->SetParam(
-                    nullptr,
-                    QUIC_PARAM_LEVEL_GLOBAL,
-                    QUIC_PARAM_GLOBAL_TEST_DATAPATH_HOOKS,
-                    sizeof(Value),
-                    &Value));
-        }
-    }
-    ~RandomLossHelper() {
-        if (LossPercentage != 0) {
-            QUIC_TEST_DATAPATH_HOOKS* Value = nullptr;
-            uint32_t TryCount = 0;
-            while (TryCount++ < 10) {
-                if (QUIC_SUCCEEDED(
-                    MsQuic->SetParam(
-                        nullptr,
-                        QUIC_PARAM_LEVEL_GLOBAL,
-                        QUIC_PARAM_GLOBAL_TEST_DATAPATH_HOOKS,
-                        sizeof(Value),
-                        &Value))) {
-                    break;
-                }
-                QuicSleep(100); // Let the current datapath queue drain.
-            }
-            if (TryCount == 10) {
-                TEST_FAILURE("Failed to disable test datapath hook");
-            }
-        }
-    }
-    static
-    _IRQL_requires_max_(DISPATCH_LEVEL)
-    BOOLEAN
-    QUIC_API
-    ReceiveCallback(
-        _Inout_ struct QUIC_RECV_DATAGRAM* /* Datagram */
-        )
-    {
-        uint8_t RandomValue;
-        QuicRandom(sizeof(RandomValue), &RandomValue);
-        return (RandomValue % 100) < LossPercentage;
-    }
-    static
-    _IRQL_requires_max_(PASSIVE_LEVEL)
-    BOOLEAN
-    QUIC_API
-    SendCallback(
-        _Inout_ QUIC_ADDR* /* RemoteAddress */,
-        _Inout_opt_ QUIC_ADDR* /* LocalAddress */,
-        _Inout_ struct QUIC_DATAPATH_SEND_CONTEXT* /* SendContext */
-        )
-    {
-        return FALSE; // Don't drop
-    }
-};
-
-uint8_t RandomLossHelper::LossPercentage = 0;
-QUIC_TEST_DATAPATH_HOOKS RandomLossHelper::DataPathFuncTable = {
-    RandomLossHelper::ReceiveCallback,
-    RandomLossHelper::SendCallback
-};
-
-void
-QuicTestConnect(
-    _In_ int Family,
-    _In_ bool ServerStatelessRetry,
-    _In_ bool ClientUsesOldVersion,
-    _In_ bool ClientRebind,
-    _In_ bool ChangeMaxStreamID,
-    _In_ bool MultipleALPNs,
-    _In_ bool AsyncSecConfig,
-    _In_ bool MultiPacketClientInitial,
-    _In_ bool SessionResumption,
-    _In_ uint8_t RandomLossPercentage
-    )
-{
-    MsQuicSession Session;
-    TEST_TRUE(Session.IsValid());
-    TEST_QUIC_SUCCEEDED(Session.SetPeerBidiStreamCount(4));
-    TEST_QUIC_SUCCEEDED(Session.SetIdleTimeout(10000));
-    MsQuicSession Session2("MsQuicTest2", "MsQuicTest");
-    TEST_TRUE(Session2.IsValid());
-    TEST_QUIC_SUCCEEDED(Session2.SetPeerBidiStreamCount(4));
-    TEST_QUIC_SUCCEEDED(Session2.SetIdleTimeout(10000));
-
-    StatelessRetryHelper RetryHelper(ServerStatelessRetry);
-    PrivateTransportHelper TpHelper(MultiPacketClientInitial);
-    RandomLossHelper LossHelper(RandomLossPercentage);
-
-    {
-        TestListener Listener(
-            MultipleALPNs ? Session2.Handle : Session.Handle,
-            ListenerAcceptConnection,
-            AsyncSecConfig);
-        TEST_TRUE(Listener.IsValid());
-        Listener.SetHasRandomLoss(RandomLossPercentage != 0);
-
-        QUIC_ADDRESS_FAMILY QuicAddrFamily = (Family == 4) ? AF_INET : AF_INET6;
-        QuicAddr ServerLocalAddr(QuicAddrFamily);
-        TEST_QUIC_SUCCEEDED(Listener.Start(&ServerLocalAddr.SockAddr));
-        TEST_QUIC_SUCCEEDED(Listener.GetLocalAddr(ServerLocalAddr));
-
-        if (SessionResumption) {
-            TestScopeLogger logScope("PrimeResumption");
-            {
-                TestConnection Client(
-                    Session.Handle,
-                    ConnectionDoNothingCallback,
-                    false);
-                TEST_TRUE(Client.IsValid());
-                Client.SetHasRandomLoss(RandomLossPercentage != 0);
-                #if QUIC_TEST_DISABLE_DNS
-                QuicAddr RemoteAddr(QuicAddrFamily, true);
-                TEST_QUIC_SUCCEEDED(Client.SetRemoteAddr(RemoteAddr));
-                #endif
-                TEST_QUIC_SUCCEEDED(
-                    Client.Start(
-                        QuicAddrFamily,
-                        QUIC_LOCALHOST_FOR_AF(QuicAddrFamily),
-                        QuicAddrGetPort(&ServerLocalAddr.SockAddr)));
-                if (!Client.WaitForConnectionComplete()) {
-                    return;
-                }
-                TEST_TRUE(Client.GetIsConnected());
-                if (!Client.WaitForZeroRttTicket()) {
-                    return;
-                }
-                Client.Shutdown(QUIC_CONNECTION_SHUTDOWN_FLAG_NONE, QUIC_TEST_NO_ERROR);
-                if (!Client.WaitForShutdownComplete()) {
-                    return;
-                }
-            }
-        }
-
-        {
-            UniquePtr<TestConnection> Server;
-            ServerAcceptContext ServerAcceptCtx((TestConnection**)&Server);
-            Listener.Context = &ServerAcceptCtx;
-
-            {
-                TestConnection Client(
-                    Session.Handle,
-                    ConnectionDoNothingCallback,
-                    false);
-                TEST_TRUE(Client.IsValid());
-                Client.SetHasRandomLoss(RandomLossPercentage != 0);
-
-                if (ClientUsesOldVersion) {
-                    TEST_QUIC_SUCCEEDED(
-                        Client.SetQuicVersion(OLD_SUPPORTED_VERSION));
-                }
-
-                if (MultiPacketClientInitial) {
-                    TEST_QUIC_SUCCEEDED(
-                        Client.SetTestTransportParameter(&TpHelper));
-                }
-
-                if (SessionResumption) {
-                    Client.SetExpectedResumed(true);
-                }
-
-                #if QUIC_TEST_DISABLE_DNS
-                QuicAddr RemoteAddr(QuicAddrFamily, true);
-                TEST_QUIC_SUCCEEDED(Client.SetRemoteAddr(RemoteAddr));
-                #endif
-
-                TEST_QUIC_SUCCEEDED(
-                    Client.Start(
-                        QuicAddrFamily,
-                        QUIC_LOCALHOST_FOR_AF(QuicAddrFamily),
-                        QuicAddrGetPort(&ServerLocalAddr.SockAddr)));
-
-                if (AsyncSecConfig) {
-                    if (!QuicEventWaitWithTimeout(ServerAcceptCtx.NewConnectionReady, TestWaitTimeout)) {
-                        TEST_FAILURE("Timed out waiting for server accept.");
-                    } else if (Server == nullptr) {
-                        TEST_FAILURE("Failed to accept server connection.");
-                    } else {
-                        TEST_QUIC_SUCCEEDED(
-                            Server->SetSecurityConfig(SecurityConfig));
-                    }
-                }
-
-                if (!Client.WaitForConnectionComplete()) {
-                    return;
-                }
-                TEST_TRUE(Client.GetIsConnected());
-
-                TEST_NOT_EQUAL(nullptr, Server);
-                if (!Server->WaitForConnectionComplete()) {
-                    return;
-                }
-                TEST_TRUE(Server->GetIsConnected());
-
-                if (ClientUsesOldVersion) {
-                    TEST_EQUAL(Server->GetQuicVersion(), OLD_SUPPORTED_VERSION);
-                } else {
-                    TEST_EQUAL(Server->GetQuicVersion(), LATEST_SUPPORTED_VERSION);
-                }
-
-                if (SessionResumption) {
-                    TEST_TRUE(Client.GetResumed());
-                    TEST_TRUE(Server->GetResumed());
-                }
-
-                TEST_EQUAL(
-                    Server->GetPeerBidiStreamCount(),
-                    Client.GetLocalBidiStreamCount());
-
-                if (ClientRebind) {
-                    QuicAddr NewLocalAddr(QuicAddrFamily);
-                    TEST_QUIC_SUCCEEDED(Client.SetLocalAddr(NewLocalAddr));
-                    QuicSleep(100);
-                    TEST_QUIC_SUCCEEDED(Client.GetLocalAddr(NewLocalAddr));
-                    TEST_FALSE(Client.GetIsShutdown());
-
-                    bool ServerAddressUpdated = false;
-                    uint32_t Try = 0;
-                    do {
-                        if (Try != 0) {
-                            QuicSleep(200);
-                        }
-                        QuicAddr ServerRemoteAddr;
-                        TEST_QUIC_SUCCEEDED(Server->GetRemoteAddr(ServerRemoteAddr));
-                        if (Server->GetPeerAddrChanged() &&
-                            QuicAddrCompare(&NewLocalAddr.SockAddr, &ServerRemoteAddr.SockAddr)) {
-                            ServerAddressUpdated = true;
-                            break;
-                        }
-                    } while (++Try <= 3);
-                    TEST_TRUE(ServerAddressUpdated);
-                }
-
-                if (ChangeMaxStreamID) {
-                    TEST_QUIC_SUCCEEDED(Client.SetPeerBidiStreamCount(101));
-                    TEST_EQUAL(101, Client.GetPeerBidiStreamCount());
-                    QuicSleep(100);
-                    TEST_EQUAL(101, Server->GetLocalBidiStreamCount());
-
-                    TEST_QUIC_SUCCEEDED(Server->SetPeerBidiStreamCount(100));
-                    TEST_EQUAL(100, Server->GetPeerBidiStreamCount());
-                    QuicSleep(100);
-                    TEST_EQUAL(100, Client.GetLocalBidiStreamCount());
-                }
-
-                Client.Shutdown(QUIC_CONNECTION_SHUTDOWN_FLAG_NONE, QUIC_TEST_NO_ERROR);
-                if (!Client.WaitForShutdownComplete()) {
-                    return;
-                }
-
-                TEST_FALSE(Client.GetPeerClosed());
-                TEST_FALSE(Client.GetTransportClosed());
-            }
-
-            if (RandomLossPercentage == 0) {
-                TEST_TRUE(Server->GetPeerClosed());
-                TEST_EQUAL(Server->GetPeerCloseErrorCode(), QUIC_TEST_NO_ERROR);
-            }
-        }
-    }
-}
 
 /*
     Helper function to estimate a maximum timeout for a test with a
@@ -1008,10 +382,6 @@ QuicTestConnectAndPing(
             {
                 TestConnection Client(Session.Handle, ConnectionDoNothingCallback, false);
                 TEST_TRUE(Client.IsValid());
-                #if QUIC_TEST_DISABLE_DNS
-                QuicAddr RemoteAddr(QuicAddrFamily, true);
-                TEST_QUIC_SUCCEEDED(Client.SetRemoteAddr(RemoteAddr));
-                #endif
                 TEST_QUIC_SUCCEEDED(
                     Client.Start(
                         QuicAddrFamily,
@@ -1098,97 +468,6 @@ QuicTestConnectAndPing(
 }
 
 void
-QuicTestConnectAndIdle(
-    _In_ bool EnableKeepAlive
-    )
-{
-    MsQuicSession Session;
-    TEST_TRUE(Session.IsValid());
-    TEST_QUIC_SUCCEEDED(Session.SetIdleTimeout(3000));
-
-    {
-        TestListener Listener(Session.Handle, ListenerAcceptConnection);
-        TEST_TRUE(Listener.IsValid());
-        TEST_QUIC_SUCCEEDED(Listener.Start());
-
-        QuicAddr ServerLocalAddr;
-        TEST_QUIC_SUCCEEDED(Listener.GetLocalAddr(ServerLocalAddr));
-
-        {
-            UniquePtr<TestConnection> Server;
-            ServerAcceptContext ServerAcceptCtx((TestConnection**)&Server);
-            Listener.Context = &ServerAcceptCtx;
-
-            {
-                TestConnection Client(
-                    Session.Handle,
-                    ConnectionDoNothingCallback,
-                    false);
-                TEST_TRUE(Client.IsValid());
-
-                if (!EnableKeepAlive) {
-                    Client.SetExpectedTransportCloseStatus(QUIC_STATUS_CONNECTION_IDLE);
-                }
-
-                #if QUIC_TEST_DISABLE_DNS
-                QuicAddr RemoteAddr(AF_INET, true);
-                TEST_QUIC_SUCCEEDED(Client.SetRemoteAddr(RemoteAddr));
-                #endif
-
-                TEST_QUIC_SUCCEEDED(
-                    Client.Start(
-                        AF_UNSPEC,
-                        QUIC_LOCALHOST_FOR_AF(ServerLocalAddr.SockAddr.si_family),
-                        QuicAddrGetPort(&ServerLocalAddr.SockAddr)));
-
-                if (!Client.WaitForConnectionComplete()) {
-                    return;
-                }
-                TEST_TRUE(Client.GetIsConnected());
-
-                TEST_NOT_EQUAL(nullptr, Server);
-                if (!EnableKeepAlive) {
-                    Server->SetExpectedTransportCloseStatus(QUIC_STATUS_CONNECTION_IDLE);
-                }
-                if (!Server->WaitForConnectionComplete()) {
-                    return;
-                }
-                TEST_TRUE(Server->GetIsConnected());
-
-                if (EnableKeepAlive) {
-                    TEST_QUIC_SUCCEEDED(Client.SetKeepAlive(1000));
-                }
-
-                QuicSleep(4000); // Wait for the first idle period to expire.
-
-                if (EnableKeepAlive) {
-                    TEST_FALSE(Client.GetIsShutdown());
-                    TEST_FALSE(Server->GetIsShutdown());
-
-                    Client.Shutdown(QUIC_CONNECTION_SHUTDOWN_FLAG_NONE, QUIC_TEST_NO_ERROR);
-                    if (!Client.WaitForShutdownComplete()) {
-                        return;
-                    }
-
-                    TEST_FALSE(Client.GetPeerClosed());
-                    TEST_FALSE(Client.GetTransportClosed());
-
-#if !QUIC_SEND_FAKE_LOSS
-                    TEST_TRUE(Server->GetPeerClosed());
-                    TEST_EQUAL(Server->GetPeerCloseErrorCode(), QUIC_TEST_NO_ERROR);
-#endif
-                } else {
-                    TEST_TRUE(Client.GetIsShutdown());
-                    TEST_TRUE(Server->GetIsShutdown());
-                    TEST_TRUE(Client.GetTransportClosed());
-                    TEST_TRUE(Server->GetTransportClosed());
-                }
-            }
-        }
-    }
-}
-
-void
 QuicTestServerDisconnect(
     void
     )
@@ -1220,11 +499,6 @@ QuicTestServerDisconnect(
                     return;
                 }
                 TEST_QUIC_SUCCEEDED(Client->SetPeerUnidiStreamCount(1));
-
-                #if QUIC_TEST_DISABLE_DNS
-                QuicAddr RemoteAddr(AF_INET, true);
-                TEST_QUIC_SUCCEEDED(Client->SetRemoteAddr(RemoteAddr));
-                #endif
 
                 TEST_QUIC_SUCCEEDED(
                     Client->Start(
@@ -1343,11 +617,6 @@ QuicTestClientDisconnect(
                     return;
                 }
 
-                #if QUIC_TEST_DISABLE_DNS
-                QuicAddr RemoteAddr(AF_INET, true);
-                TEST_QUIC_SUCCEEDED(Client->SetRemoteAddr(RemoteAddr));
-                #endif
-
                 TEST_QUIC_SUCCEEDED(
                     Client->Start(
                         AF_INET,
@@ -1377,395 +646,6 @@ QuicTestClientDisconnect(
             (void)Client->WaitForShutdownComplete();
         }
     } // Scope exit waits on Session closure, which waits for connection closures.
-}
-
-void
-QuicTestConnectUnreachable(
-    _In_ int Family
-    )
-{
-    MsQuicSession Session;
-    TEST_TRUE(Session.IsValid());
-
-    {
-        QUIC_ADDRESS_FAMILY QuicAddrFamily = (Family == 4) ? AF_INET : AF_INET6;
-
-        TestConnection Client(
-            Session.Handle,
-            ConnectionDoNothingCallback,
-            false);
-        TEST_TRUE(Client.IsValid());
-
-        #if QUIC_TEST_DISABLE_DNS
-        QuicAddr RemoteAddr(Family == 4 ? AF_INET : AF_INET6, true);
-        TEST_QUIC_SUCCEEDED(Client.SetRemoteAddr(RemoteAddr));
-        #endif
-
-        Client.SetExpectedTransportCloseStatus(QUIC_STATUS_UNREACHABLE);
-        TEST_QUIC_SUCCEEDED(
-            Client.Start(
-                QuicAddrFamily,
-                QUIC_LOCALHOST_FOR_AF(QuicAddrFamily),
-                TestUdpPortBase - 1));
-        if (!Client.WaitForConnectionComplete()) {
-            return;
-        }
-
-        TEST_FALSE(Client.GetIsConnected());
-        TEST_TRUE(Client.GetTransportClosed());
-    }
-}
-
-void
-QuicTestVersionNegotiation(
-    _In_ int Family
-    )
-{
-    MsQuicSession Session;
-    TEST_TRUE(Session.IsValid());
-    TEST_QUIC_SUCCEEDED(Session.SetIdleTimeout(3000));
-
-    {
-        TestListener Listener(Session.Handle, ListenerAcceptConnection);
-        TEST_TRUE(Listener.IsValid());
-        TEST_QUIC_SUCCEEDED(Listener.Start());
-
-        QUIC_ADDRESS_FAMILY QuicAddrFamily = (Family == 4) ? AF_INET : AF_INET6;
-        QuicAddr ServerLocalAddr;
-        TEST_QUIC_SUCCEEDED(Listener.GetLocalAddr(ServerLocalAddr));
-
-        {
-            UniquePtr<TestConnection> Server;
-            ServerAcceptContext ServerAcceptCtx((TestConnection**)&Server);
-            Listener.Context = &ServerAcceptCtx;
-
-            {
-                TestConnection Client(
-                    Session.Handle,
-                    ConnectionDoNothingCallback,
-                    false);
-                TEST_TRUE(Client.IsValid());
-
-                #if QUIC_TEST_DISABLE_DNS
-                QuicAddr RemoteAddr(QuicAddrFamily, true);
-                TEST_QUIC_SUCCEEDED(Client.SetRemoteAddr(RemoteAddr));
-                #endif
-
-                TEST_QUIC_SUCCEEDED(
-                    Client.SetQuicVersion(168430090ul)); // Random reserved version to force VN.
-
-                Client.SetExpectedTransportCloseStatus(QUIC_STATUS_VER_NEG_ERROR);
-                TEST_QUIC_SUCCEEDED(
-                    Client.Start(
-                        QuicAddrFamily,
-                        QUIC_LOCALHOST_FOR_AF(QuicAddrFamily),
-                        QuicAddrGetPort(&ServerLocalAddr.SockAddr)));
-                if (!Client.WaitForConnectionComplete()) {
-                    return;
-                }
-
-                TEST_FALSE(Client.GetIsConnected());
-                TEST_TRUE(Client.GetTransportClosed());
-
-                TEST_EQUAL(nullptr, Server);
-            }
-        }
-    }
-}
-
-void
-QuicTestConnectBadAlpn(
-    _In_ int Family
-    )
-{
-    MsQuicSession GoodSession;
-    TEST_TRUE(GoodSession.IsValid());
-    TEST_QUIC_SUCCEEDED(GoodSession.SetIdleTimeout(3000));
-    MsQuicSession BadSession("BadALPN"); // Incorrect ALPN
-    TEST_TRUE(BadSession.IsValid());
-
-    {
-        TestListener Listener(GoodSession.Handle, ListenerAcceptConnection);
-        TEST_TRUE(Listener.IsValid());
-        TEST_QUIC_SUCCEEDED(Listener.Start());
-
-        QUIC_ADDRESS_FAMILY QuicAddrFamily = (Family == 4) ? AF_INET : AF_INET6;
-        QuicAddr ServerLocalAddr;
-        TEST_QUIC_SUCCEEDED(Listener.GetLocalAddr(ServerLocalAddr));
-
-        {
-            UniquePtr<TestConnection> Server;
-            ServerAcceptContext ServerAcceptCtx((TestConnection**)&Server);
-            Listener.Context = &ServerAcceptCtx;
-
-            {
-                TestConnection Client(
-                    BadSession.Handle,
-                    ConnectionDoNothingCallback,
-                    false);
-                TEST_TRUE(Client.IsValid());
-
-                #if QUIC_TEST_DISABLE_DNS
-                QuicAddr RemoteAddr(QuicAddrFamily, true);
-                TEST_QUIC_SUCCEEDED(Client.SetRemoteAddr(RemoteAddr));
-                #endif
-
-                Client.SetExpectedTransportCloseStatus(QUIC_STATUS_INTERNAL_ERROR);
-                TEST_QUIC_SUCCEEDED(
-                    Client.Start(
-                        QuicAddrFamily,
-                        QUIC_LOCALHOST_FOR_AF(QuicAddrFamily),
-                        QuicAddrGetPort(&ServerLocalAddr.SockAddr)));
-                if (!Client.WaitForConnectionComplete()) {
-                    return;
-                }
-
-                TEST_FALSE(Client.GetIsConnected());
-                TEST_TRUE(Client.GetTransportClosed());
-
-                TEST_EQUAL(nullptr, Server);
-            }
-        }
-    }
-}
-
-void
-QuicTestConnectBadSni(
-    _In_ int Family
-    )
-{
-    MsQuicSession Session;
-    TEST_TRUE(Session.IsValid());
-    TEST_QUIC_SUCCEEDED(Session.SetIdleTimeout(3000));
-
-    {
-        TestListener Listener(Session.Handle, ListenerAcceptConnection);
-        TEST_TRUE(Listener.IsValid());
-        TEST_QUIC_SUCCEEDED(Listener.Start());
-
-        QuicAddr ServerLocalAddr;
-        TEST_QUIC_SUCCEEDED(Listener.GetLocalAddr(ServerLocalAddr));
-
-        {
-            UniquePtr<TestConnection> Server;
-            ServerAcceptContext ServerAcceptCtx((TestConnection**)&Server);
-            Listener.Context = &ServerAcceptCtx;
-
-            {
-                TestConnection Client(
-                    Session.Handle,
-                    ConnectionDoNothingCallback,
-                    false);
-                TEST_TRUE(Client.IsValid());
-
-                QuicAddr RemoteAddr(Family == 4 ? AF_INET : AF_INET6, true);
-                TEST_QUIC_SUCCEEDED(Client.SetRemoteAddr(RemoteAddr));
-
-                Client.SetExpectedTransportCloseStatus(QUIC_STATUS_INTERNAL_ERROR);
-                TEST_QUIC_SUCCEEDED(
-                    Client.Start(
-                        Family == 4 ? AF_INET : AF_INET6,
-                        "badlocalhost",
-                        QuicAddrGetPort(&ServerLocalAddr.SockAddr)));
-                if (!Client.WaitForConnectionComplete()) {
-                    return;
-                }
-
-                TEST_FALSE(Client.GetIsConnected());
-                TEST_TRUE(Client.GetTransportClosed());
-
-                TEST_EQUAL(nullptr, Server);
-            }
-        }
-    }
-}
-
-_Function_class_(NEW_CONNECTION_CALLBACK)
-static
-void
-ListenerRejectConnection(
-    _In_ TestListener* /*  Listener */,
-    _In_ HQUIC ConnectionHandle
-    )
-{
-    auto Connection = new TestConnection(ConnectionHandle, ConnectionDoNothingCallback, true, true);
-    Connection->Shutdown(QUIC_CONNECTION_SHUTDOWN_FLAG_NONE, QUIC_TEST_SPECIAL_ERROR);
-}
-
-void
-QuicTestConnectServerRejected(
-    _In_ int Family
-    )
-{
-    MsQuicSession Session;
-    TEST_TRUE(Session.IsValid());
-    TEST_QUIC_SUCCEEDED(Session.SetIdleTimeout(3000));
-
-    {
-        TestListener Listener(Session.Handle, ListenerRejectConnection, true);
-        TEST_TRUE(Listener.IsValid());
-        TEST_QUIC_SUCCEEDED(Listener.Start());
-
-        QUIC_ADDRESS_FAMILY QuicAddrFamily = (Family == 4) ? AF_INET : AF_INET6;
-        QuicAddr ServerLocalAddr;
-        TEST_QUIC_SUCCEEDED(Listener.GetLocalAddr(ServerLocalAddr));
-
-        {
-            TestConnection Client(
-                Session.Handle,
-                ConnectionDoNothingCallback,
-                false);
-            TEST_TRUE(Client.IsValid());
-
-            #if QUIC_TEST_DISABLE_DNS
-            QuicAddr RemoteAddr(QuicAddrFamily, true);
-            TEST_QUIC_SUCCEEDED(Client.SetRemoteAddr(RemoteAddr));
-            #endif
-
-            Client.SetExpectedTransportCloseStatus(QUIC_STATUS_USER_CANCELED);
-            TEST_QUIC_SUCCEEDED(
-                Client.Start(
-                    QuicAddrFamily,
-                    QUIC_LOCALHOST_FOR_AF(QuicAddrFamily),
-                    QuicAddrGetPort(&ServerLocalAddr.SockAddr)));
-            if (!Client.WaitForConnectionComplete()) {
-                return;
-            }
-
-            TEST_FALSE(Client.GetIsConnected());
-            TEST_TRUE(Client.GetTransportClosed());
-        }
-    }
-}
-
-void
-QuicTestKeyUpdate(
-    _In_ int Family,
-    _In_ uint16_t Iterations,
-    _In_ uint16_t KeyUpdateBytes,
-    _In_ bool UseKeyUpdateBytes,
-    _In_ bool ClientKeyUpdate,
-    _In_ bool ServerKeyUpdate
-    )
-{
-    MsQuicSession Session;
-    TEST_TRUE(Session.IsValid());
-
-    if (UseKeyUpdateBytes) {
-        Session.SetMaxBytesPerKey((uint64_t)KeyUpdateBytes);
-    }
-
-    {
-        TestListener Listener(Session.Handle, ListenerAcceptConnection);
-        TEST_TRUE(Listener.IsValid());
-
-        QUIC_ADDRESS_FAMILY QuicAddrFamily = (Family == 4) ? AF_INET : AF_INET6;
-        QuicAddr ServerLocalAddr(QuicAddrFamily);
-        TEST_QUIC_SUCCEEDED(Listener.Start(&ServerLocalAddr.SockAddr));
-        TEST_QUIC_SUCCEEDED(Listener.GetLocalAddr(ServerLocalAddr));
-
-        {
-            UniquePtr<TestConnection> Server;
-            ServerAcceptContext ServerAcceptCtx((TestConnection**)&Server);
-            Listener.Context = &ServerAcceptCtx;
-
-            {
-                TestConnection Client(
-                    Session.Handle,
-                    ConnectionDoNothingCallback,
-                    false);
-                TEST_TRUE(Client.IsValid());
-
-                #if QUIC_TEST_DISABLE_DNS
-                QuicAddr RemoteAddr(QuicAddrFamily, true);
-                TEST_QUIC_SUCCEEDED(Client.SetRemoteAddr(RemoteAddr));
-                #endif
-
-                TEST_QUIC_SUCCEEDED(
-                    Client.Start(
-                        QuicAddrFamily,
-                        QUIC_LOCALHOST_FOR_AF(QuicAddrFamily),
-                        QuicAddrGetPort(&ServerLocalAddr.SockAddr)));
-
-                if (!Client.WaitForConnectionComplete()) {
-                    return;
-                }
-                TEST_TRUE(Client.GetIsConnected());
-
-                TEST_NOT_EQUAL(nullptr, Server);
-                if (!Server->WaitForConnectionComplete()) {
-                    return;
-                }
-                TEST_TRUE(Server->GetIsConnected());
-
-                for (uint16_t i = 0; i < Iterations; ++i) {
-
-                    QuicSleep(100);
-
-                    if (ClientKeyUpdate) {
-                        TEST_QUIC_SUCCEEDED(Client.ForceKeyUpdate());
-                    }
-
-                    if (ServerKeyUpdate) {
-                        TEST_QUIC_SUCCEEDED(Server->ForceKeyUpdate());
-                    }
-
-                    //
-                    // Send some data to perform the key update.
-                    // TODO: Update this to send stream data, like QuicConnectAndPing does.
-                    //
-                    TEST_QUIC_SUCCEEDED(Client.SetPeerBidiStreamCount((uint16_t)(101+i)));
-                    TEST_EQUAL((uint16_t)(101+i), Client.GetPeerBidiStreamCount());
-                    QuicSleep(100);
-                    TEST_EQUAL((uint16_t)(101+i), Server->GetLocalBidiStreamCount());
-
-                    TEST_QUIC_SUCCEEDED(Server->SetPeerBidiStreamCount((uint16_t)(100+i)));
-                    TEST_EQUAL((uint16_t)(100+i), Server->GetPeerBidiStreamCount());
-                    QuicSleep(100);
-                    TEST_EQUAL((uint16_t)(100+i), Client.GetLocalBidiStreamCount());
-                }
-
-                QuicSleep(100);
-
-                QUIC_STATISTICS Stats = Client.GetStatistics();
-                if (Stats.Recv.DecryptionFailures) {
-                    TEST_FAILURE("%llu server packets failed to decrypt!", Stats.Recv.DecryptionFailures);
-                    return;
-                }
-
-                uint16_t ExpectedUpdates = Iterations - (UseKeyUpdateBytes ? 1u : 0u);
-
-                if (Stats.Misc.KeyUpdateCount < ExpectedUpdates) {
-                    TEST_FAILURE("%u Key updates occured. Expected %d", Stats.Misc.KeyUpdateCount, ExpectedUpdates);
-                    return;
-                }
-
-                Stats = Server->GetStatistics();
-                if (Stats.Recv.DecryptionFailures) {
-                    TEST_FAILURE("%llu client packets failed to decrypt!", Stats.Recv.DecryptionFailures);
-                    return;
-                }
-
-                if (Stats.Misc.KeyUpdateCount < ExpectedUpdates) {
-                    TEST_FAILURE("%u Key updates occured. Expected %d", Stats.Misc.KeyUpdateCount, ExpectedUpdates);
-                    return;
-                }
-
-                Client.Shutdown(QUIC_CONNECTION_SHUTDOWN_FLAG_NONE, QUIC_TEST_NO_ERROR);
-                if (!Client.WaitForShutdownComplete()) {
-                    return;
-                }
-
-                TEST_FALSE(Client.GetPeerClosed());
-                TEST_FALSE(Client.GetTransportClosed());
-            }
-
-#if !QUIC_SEND_FAKE_LOSS
-            TEST_TRUE(Server->GetPeerClosed());
-            TEST_EQUAL(Server->GetPeerCloseErrorCode(), QUIC_TEST_NO_ERROR);
-#endif
-        }
-    }
 }
 
 struct AbortiveTestContext {
@@ -2236,106 +1116,6 @@ QuicAbortiveTransfers(
             }
             TEST_EQUAL(ExpectedError, ServerContext.TestResult);
             TEST_TRUE(ServerContext.Passed);
-        }
-    }
-}
-
-void
-QuicTestCidUpdate(
-    _In_ int Family,
-    _In_ uint16_t Iterations
-    )
-{
-    MsQuicSession Session;
-    TEST_TRUE(Session.IsValid());
-
-    {
-        TestListener Listener(Session.Handle, ListenerAcceptConnection);
-        TEST_TRUE(Listener.IsValid());
-
-        QUIC_ADDRESS_FAMILY QuicAddrFamily = (Family == 4) ? AF_INET : AF_INET6;
-        QuicAddr ServerLocalAddr(QuicAddrFamily);
-        TEST_QUIC_SUCCEEDED(Listener.Start(&ServerLocalAddr.SockAddr));
-        TEST_QUIC_SUCCEEDED(Listener.GetLocalAddr(ServerLocalAddr));
-
-        {
-            UniquePtr<TestConnection> Server;
-            ServerAcceptContext ServerAcceptCtx((TestConnection**)&Server);
-            Listener.Context = &ServerAcceptCtx;
-
-            {
-                TestConnection Client(
-                    Session.Handle,
-                    ConnectionDoNothingCallback,
-                    false);
-                TEST_TRUE(Client.IsValid());
-
-                #if QUIC_TEST_DISABLE_DNS
-                QuicAddr RemoteAddr(QuicAddrFamily, true);
-                TEST_QUIC_SUCCEEDED(Client.SetRemoteAddr(RemoteAddr));
-                #endif
-
-                TEST_QUIC_SUCCEEDED(
-                    Client.Start(
-                        QuicAddrFamily,
-                        QUIC_LOCALHOST_FOR_AF(QuicAddrFamily),
-                        QuicAddrGetPort(&ServerLocalAddr.SockAddr)));
-
-                if (!Client.WaitForConnectionComplete()) {
-                    return;
-                }
-                TEST_TRUE(Client.GetIsConnected());
-
-                TEST_NOT_EQUAL(nullptr, Server);
-                if (!Server->WaitForConnectionComplete()) {
-                    return;
-                }
-                TEST_TRUE(Server->GetIsConnected());
-
-                for (uint16_t i = 0; i < Iterations; ++i) {
-
-                    QuicSleep(100);
-
-                    TEST_QUIC_SUCCEEDED(Client.ForceCidUpdate());
-
-                    TEST_QUIC_SUCCEEDED(Client.SetPeerBidiStreamCount((uint16_t)(101+i)));
-                    TEST_EQUAL((uint16_t)(101+i), Client.GetPeerBidiStreamCount());
-                    QuicSleep(100);
-                    TEST_EQUAL((uint16_t)(101+i), Server->GetLocalBidiStreamCount());
-
-                    TEST_QUIC_SUCCEEDED(Server->SetPeerBidiStreamCount((uint16_t)(100+i)));
-                    TEST_EQUAL((uint16_t)(100+i), Server->GetPeerBidiStreamCount());
-                    QuicSleep(100);
-                    TEST_EQUAL((uint16_t)(100+i), Client.GetLocalBidiStreamCount());
-                }
-
-                QuicSleep(100);
-
-                QUIC_STATISTICS Stats = Client.GetStatistics();
-                if (Stats.Recv.DecryptionFailures) {
-                    TEST_FAILURE("%llu server packets failed to decrypt!", Stats.Recv.DecryptionFailures);
-                    return;
-                }
-
-                Stats = Server->GetStatistics();
-                if (Stats.Recv.DecryptionFailures) {
-                    TEST_FAILURE("%llu client packets failed to decrypt!", Stats.Recv.DecryptionFailures);
-                    return;
-                }
-
-                Client.Shutdown(QUIC_CONNECTION_SHUTDOWN_FLAG_NONE, QUIC_TEST_NO_ERROR);
-                if (!Client.WaitForShutdownComplete()) {
-                    return;
-                }
-
-                TEST_FALSE(Client.GetPeerClosed());
-                TEST_FALSE(Client.GetTransportClosed());
-            }
-
-#if !QUIC_SEND_FAKE_LOSS
-            TEST_TRUE(Server->GetPeerClosed());
-            TEST_EQUAL(Server->GetPeerCloseErrorCode(), QUIC_TEST_NO_ERROR);
-#endif
         }
     }
 }
@@ -2985,291 +1765,6 @@ QuicTestReceiveResumeNoData(
         if (!QuicEventWaitWithTimeout(ServerContext.TestEvent.Handle, TimeoutMs)) {
             TEST_FAILURE("Server failed to get shutdown before timeout!");
             return;
-        }
-    }
-}
-
-void
-QuicTestDatagramNegotiation(
-    _In_ int Family,
-    _In_ bool DatagramReceiveEnabled
-    )
-{
-    MsQuicSession ClientSession;
-    TEST_TRUE(ClientSession.IsValid());
-    TEST_QUIC_SUCCEEDED(ClientSession.SetDatagramReceiveEnabled(true)); // Always enabled on client.
-
-    MsQuicSession ServerSession;
-    TEST_TRUE(ServerSession.IsValid());
-    TEST_QUIC_SUCCEEDED(ServerSession.SetDatagramReceiveEnabled(DatagramReceiveEnabled));
-
-    uint8_t RawBuffer[] = "datagram";
-    QUIC_BUFFER DatagramBuffer = { sizeof(RawBuffer), RawBuffer };
-
-    {
-        TestListener Listener(ServerSession.Handle, ListenerAcceptConnection);
-        TEST_TRUE(Listener.IsValid());
-
-        QUIC_ADDRESS_FAMILY QuicAddrFamily = (Family == 4) ? AF_INET : AF_INET6;
-        QuicAddr ServerLocalAddr(QuicAddrFamily);
-        TEST_QUIC_SUCCEEDED(Listener.Start(&ServerLocalAddr.SockAddr));
-        TEST_QUIC_SUCCEEDED(Listener.GetLocalAddr(ServerLocalAddr));
-
-        {
-            UniquePtr<TestConnection> Server;
-            ServerAcceptContext ServerAcceptCtx((TestConnection**)&Server);
-            Listener.Context = &ServerAcceptCtx;
-
-            {
-                TestConnection Client(
-                    ClientSession.Handle,
-                    ConnectionDoNothingCallback,
-                    false);
-                TEST_TRUE(Client.IsValid());
-
-                TEST_TRUE(Client.GetDatagramSendEnabled()); // Datagrams start as enabled
-
-                TEST_QUIC_SUCCEEDED(
-                    MsQuic->DatagramSend(
-                        Client.GetConnection(),
-                        &DatagramBuffer,
-                        1,
-                        QUIC_SEND_FLAG_NONE,
-                        nullptr));
-
-                #if QUIC_TEST_DISABLE_DNS
-                QuicAddr RemoteAddr(QuicAddrFamily, true);
-                TEST_QUIC_SUCCEEDED(Client.SetRemoteAddr(RemoteAddr));
-                #endif
-
-                TEST_QUIC_SUCCEEDED(
-                    Client.Start(
-                        QuicAddrFamily,
-                        QUIC_LOCALHOST_FOR_AF(QuicAddrFamily),
-                        QuicAddrGetPort(&ServerLocalAddr.SockAddr)));
-
-                if (!Client.WaitForConnectionComplete()) {
-                    return;
-                }
-                TEST_TRUE(Client.GetIsConnected());
-
-                TEST_EQUAL(DatagramReceiveEnabled, Client.GetDatagramSendEnabled());
-
-                TEST_NOT_EQUAL(nullptr, Server);
-                if (!Server->WaitForConnectionComplete()) {
-                    return;
-                }
-                TEST_TRUE(Server->GetIsConnected());
-
-                TEST_TRUE(Server->GetDatagramSendEnabled()); // Client always enabled
-
-                QuicSleep(100); // Necessary?
-
-                if (DatagramReceiveEnabled) {
-                    TEST_EQUAL(1, Client.GetDatagramsSent());
-                } else {
-                    TEST_EQUAL(1, Client.GetDatagramsCanceled());
-                }
-
-                Client.Shutdown(QUIC_CONNECTION_SHUTDOWN_FLAG_NONE, QUIC_TEST_NO_ERROR);
-                if (!Client.WaitForShutdownComplete()) {
-                    return;
-                }
-
-                TEST_FALSE(Client.GetPeerClosed());
-                TEST_FALSE(Client.GetTransportClosed());
-            }
-
-#if !QUIC_SEND_FAKE_LOSS
-            TEST_TRUE(Server->GetPeerClosed());
-            TEST_EQUAL(Server->GetPeerCloseErrorCode(), QUIC_TEST_NO_ERROR);
-#endif
-        }
-    }
-}
-
-struct SelectiveLossHelper
-{
-    static uint32_t DropPacketCount;
-    static QUIC_TEST_DATAPATH_HOOKS DataPathFuncTable;
-    SelectiveLossHelper() {
-        QUIC_TEST_DATAPATH_HOOKS* Value = &DataPathFuncTable;
-        TEST_QUIC_SUCCEEDED(
-            MsQuic->SetParam(
-                nullptr,
-                QUIC_PARAM_LEVEL_GLOBAL,
-                QUIC_PARAM_GLOBAL_TEST_DATAPATH_HOOKS,
-                sizeof(Value),
-                &Value));
-    }
-    ~SelectiveLossHelper() {
-        QUIC_TEST_DATAPATH_HOOKS* Value = nullptr;
-        uint32_t TryCount = 0;
-        while (TryCount++ < 10) {
-            if (QUIC_SUCCEEDED(
-                MsQuic->SetParam(
-                    nullptr,
-                    QUIC_PARAM_LEVEL_GLOBAL,
-                    QUIC_PARAM_GLOBAL_TEST_DATAPATH_HOOKS,
-                    sizeof(Value),
-                    &Value))) {
-                break;
-            }
-            QuicSleep(100); // Let the current datapath queue drain.
-        }
-        if (TryCount == 10) {
-            TEST_FAILURE("Failed to disable test datapath hook");
-        }
-    }
-    void DropPackets(uint32_t Count) { DropPacketCount = Count; }
-    static
-    _IRQL_requires_max_(DISPATCH_LEVEL)
-    BOOLEAN
-    QUIC_API
-    ReceiveCallback(
-        _Inout_ struct QUIC_RECV_DATAGRAM* /* Datagram */
-        )
-    {
-        if (DropPacketCount == 0) {
-            return false;
-        }
-        DropPacketCount--;
-        return true;
-    }
-    static
-    _IRQL_requires_max_(PASSIVE_LEVEL)
-    BOOLEAN
-    QUIC_API
-    SendCallback(
-        _Inout_ QUIC_ADDR* /* RemoteAddress */,
-        _Inout_opt_ QUIC_ADDR* /* LocalAddress */,
-        _Inout_ struct QUIC_DATAPATH_SEND_CONTEXT* /* SendContext */
-        )
-    {
-        return FALSE; // Don't drop
-    }
-};
-
-uint32_t SelectiveLossHelper::DropPacketCount = false;
-QUIC_TEST_DATAPATH_HOOKS SelectiveLossHelper::DataPathFuncTable = {
-    SelectiveLossHelper::ReceiveCallback,
-    SelectiveLossHelper::SendCallback
-};
-
-void
-QuicTestDatagramSend(
-    _In_ int Family
-    )
-{
-    MsQuicSession Session;
-    TEST_TRUE(Session.IsValid());
-    TEST_QUIC_SUCCEEDED(Session.SetDatagramReceiveEnabled(true));
-
-    uint8_t RawBuffer[] = "datagram";
-    QUIC_BUFFER DatagramBuffer = { sizeof(RawBuffer), RawBuffer };
-
-    SelectiveLossHelper LossHelper;
-
-    {
-        TestListener Listener(Session.Handle, ListenerAcceptConnection);
-        TEST_TRUE(Listener.IsValid());
-
-        QUIC_ADDRESS_FAMILY QuicAddrFamily = (Family == 4) ? AF_INET : AF_INET6;
-        QuicAddr ServerLocalAddr(QuicAddrFamily);
-        TEST_QUIC_SUCCEEDED(Listener.Start(&ServerLocalAddr.SockAddr));
-        TEST_QUIC_SUCCEEDED(Listener.GetLocalAddr(ServerLocalAddr));
-
-        {
-            UniquePtr<TestConnection> Server;
-            ServerAcceptContext ServerAcceptCtx((TestConnection**)&Server);
-            Listener.Context = &ServerAcceptCtx;
-
-            {
-                TestConnection Client(
-                    Session.Handle,
-                    ConnectionDoNothingCallback,
-                    false);
-                TEST_TRUE(Client.IsValid());
-
-                TEST_TRUE(Client.GetDatagramSendEnabled());
-
-                #if QUIC_TEST_DISABLE_DNS
-                QuicAddr RemoteAddr(QuicAddrFamily, true);
-                TEST_QUIC_SUCCEEDED(Client.SetRemoteAddr(RemoteAddr));
-                #endif
-
-                TEST_QUIC_SUCCEEDED(
-                    Client.Start(
-                        QuicAddrFamily,
-                        QUIC_LOCALHOST_FOR_AF(QuicAddrFamily),
-                        QuicAddrGetPort(&ServerLocalAddr.SockAddr)));
-
-                if (!Client.WaitForConnectionComplete()) {
-                    return;
-                }
-                TEST_TRUE(Client.GetIsConnected());
-
-                TEST_TRUE(Client.GetDatagramSendEnabled());
-
-                TEST_NOT_EQUAL(nullptr, Server);
-                if (!Server->WaitForConnectionComplete()) {
-                    return;
-                }
-                TEST_TRUE(Server->GetIsConnected());
-
-                TEST_TRUE(Server->GetDatagramSendEnabled());
-
-                QuicSleep(100);
-
-                TEST_QUIC_SUCCEEDED(
-                    MsQuic->DatagramSend(
-                        Client.GetConnection(),
-                        &DatagramBuffer,
-                        1,
-                        QUIC_SEND_FLAG_NONE,
-                        nullptr));
-
-                QuicSleep(100);
-
-                TEST_EQUAL(1, Client.GetDatagramsSent());
-
-                QuicSleep(100);
-
-                TEST_EQUAL(1, Client.GetDatagramsAcknowledged());
-
-                LossHelper.DropPackets(1);
-
-                TEST_QUIC_SUCCEEDED(
-                    MsQuic->DatagramSend(
-                        Client.GetConnection(),
-                        &DatagramBuffer,
-                        1,
-                        QUIC_SEND_FLAG_NONE,
-                        nullptr));
-
-                QuicSleep(100);
-
-                TEST_EQUAL(2, Client.GetDatagramsSent());
-
-                QuicSleep(500);
-
-                TEST_EQUAL(1, Client.GetDatagramsSuspectLost());
-
-                Client.Shutdown(QUIC_CONNECTION_SHUTDOWN_FLAG_NONE, QUIC_TEST_NO_ERROR);
-                if (!Client.WaitForShutdownComplete()) {
-                    return;
-                }
-
-                TEST_EQUAL(1, Client.GetDatagramsLost());
-
-                TEST_FALSE(Client.GetPeerClosed());
-                TEST_FALSE(Client.GetTransportClosed());
-            }
-
-#if !QUIC_SEND_FAKE_LOSS
-            TEST_TRUE(Server->GetPeerClosed());
-            TEST_EQUAL(Server->GetPeerCloseErrorCode(), QUIC_TEST_NO_ERROR);
-#endif
         }
     }
 }
